# Update: Refactors the traversing logic with the decorator pattern