--- a/lib/eslint.js
+++ b/lib/eslint.js
@@ -20,6 +20,8 @@ var estraverse = require("estraverse-fb"),
     RuleContext = require("./rule-context"),
     timing = require("./timing"),
     SourceCode = require("./util/source-code"),
+    NodeEventGenerator = require("./util/node-event-generator"),
+    CommentEventGenerator = require("./util/comment-event-generator"),
     EventEmitter = require("events").EventEmitter,
     validator = require("./config-validator"),
     replacements = require("../conf/replacements.json");
@@ -498,8 +500,6 @@ module.exports = (function() {
         currentFilename = null,
         controller = null,
         reportingConfig = [],
-        commentLocsEnter = [],
-        commentLocsExit = [],
         sourceCode = null;
 
     /**
@@ -565,46 +565,6 @@ module.exports = (function() {
     }
 
     /**
-     * Check collection of comments to prevent double event for comment as
-     * leading and trailing, then emit event if passing
-     * @param {ASTNode[]} comments Collection of comment nodes
-     * @param {Object[]} locs List of locations of previous comment nodes
-     * @param {string} eventName Event name postfix
-     * @returns {void}
-     */
-    function emitComments(comments, locs, eventName) {
-
-        if (comments.length) {
-            comments.forEach(function(node) {
-                if (locs.indexOf(node.loc) >= 0) {
-                    locs.splice(locs.indexOf(node.loc), 1);
-                } else {
-                    locs.push(node.loc);
-                    api.emit(node.type + eventName, node);
-                }
-            });
-        }
-    }
-
-    /**
-     * Shortcut to check and emit enter of comment nodes
-     * @param {ASTNode[]} comments Collection of comment nodes
-     * @returns {void}
-     */
-    function emitCommentsEnter(comments) {
-        emitComments(comments, commentLocsEnter, "Comment");
-    }
-
-    /**
-     * Shortcut to check and emit exit of comment nodes
-     * @param {ASTNode[]} comments Collection of comment nodes
-     * @returns {void}
-     */
-    function emitCommentsExit(comments) {
-        emitComments(comments, commentLocsExit, "Comment:exit");
-    }
-
-    /**
      * Get the severity level of a rule (0 - none, 1 - warning, 2 - error)
      * Returns 0 if the rule config is not valid (an Array or a number)
      * @param {Array|number} ruleConfig rule configuration
@@ -649,8 +609,6 @@ module.exports = (function() {
         scopeManager = null;
         controller = null;
         reportingConfig = [];
-        commentLocsEnter = [];
-        commentLocsExit = [];
         sourceCode = null;
     };
 
@@ -811,6 +769,9 @@ module.exports = (function() {
                 }
             }
 
+            var eventGenerator = new NodeEventGenerator(api);
+            eventGenerator = new CommentEventGenerator(eventGenerator, sourceCode);
+
             /*
              * Each node has a type property. Whenever a particular type of node is found,
              * an event is fired. This allows any listeners to automatically be informed
@@ -818,21 +779,11 @@ module.exports = (function() {
              */
             controller.traverse(ast, {
                 enter: function(node, parent) {
-
-                    var comments = api.getComments(node);
-
-                    emitCommentsEnter(comments.leading);
                     node.parent = parent;
-                    api.emit(node.type, node);
-                    emitCommentsEnter(comments.trailing);
+                    eventGenerator.enterNode(node);
                 },
                 leave: function(node) {
-
-                    var comments = api.getComments(node);
-
-                    emitCommentsExit(comments.trailing);
-                    api.emit(node.type + ":exit", node);
-                    emitCommentsExit(comments.leading);
+                    eventGenerator.leaveNode(node);
                 }
             });