--- a/tests/lib/util/comment-event-generator.js
+++ b/tests/lib/util/comment-event-generator.js
@@ -1,0 +1,81 @@
+/**
+ * @fileoverview Tests for CommentEventGenerator.
+ * @author Toru Nagashima
+ * @copyright 2015 Toru Nagashima. All rights reserved.
+ * See LICENSE file in root directory for full license.
+ */
+"use strict";
+
+//------------------------------------------------------------------------------
+// Requirements
+//------------------------------------------------------------------------------
+
+var assert = require("assert"),
+    EventEmitter = require("events").EventEmitter,
+    sinon = require("sinon"),
+    espree = require("espree"),
+    estraverse = require("estraverse"),
+    EventGeneratorTester = require("../../../lib/testers/event-generator-tester"),
+    SourceCode = require("../../../lib/util/source-code"),
+    NodeEventGenerator = require("../../../lib/util/node-event-generator"),
+    CommentEventGenerator = require("../../../lib/util/comment-event-generator");
+
+//------------------------------------------------------------------------------
+// Tests
+//------------------------------------------------------------------------------
+
+describe("NodeEventGenerator", function() {
+    EventGeneratorTester.testEventGeneratorInterface(
+        new CommentEventGenerator(new NodeEventGenerator(new EventEmitter()))
+    );
+
+    it("should generate comment events without duplicate.", function() {
+        var emitter = new EventEmitter();
+        var generator = new NodeEventGenerator(emitter);
+        var code = "//foo\nvar zzz /*aaa*/ = 777\n//bar";
+        var ast = espree.parse(code, {
+            range: true,
+            loc: true,
+            comments: true,
+            attachComment: true,
+            tokens: true
+        });
+        var sourceCode = new SourceCode(code, ast);
+        var expected = [
+            ["Program", ast],
+            ["LineComment", ast.comments[0]], // foo
+            ["VariableDeclaration", ast.body[0]],
+            ["VariableDeclarator", ast.body[0].declarations[0]],
+            ["LineComment", ast.comments[2]], // bar
+            ["Identifier", ast.body[0].declarations[0].id],
+            ["BlockComment", ast.comments[1]], /* aaa */
+            ["BlockComment:exit", ast.comments[1]], /* aaa */
+            ["Identifier:exit", ast.body[0].declarations[0].id],
+            ["Literal", ast.body[0].declarations[0].init],
+            ["Literal:exit", ast.body[0].declarations[0].init],
+            ["LineComment:exit", ast.comments[0]], // foo
+            ["LineComment:exit", ast.comments[2]], // bar
+            ["VariableDeclarator:exit", ast.body[0].declarations[0]],
+            ["VariableDeclaration:exit", ast.body[0]],
+            ["Program:exit", ast]
+        ];
+
+        emitter.emit = sinon.spy(emitter.emit);
+        generator = new CommentEventGenerator(generator, sourceCode);
+
+        estraverse.traverse(ast, {
+            enter: function(node) {
+                generator.enterNode(node);
+            },
+            leave: function(node) {
+                generator.leaveNode(node);
+            }
+        });
+
+        assert.equal(emitter.emit.callCount, expected.length);
+        for (var i = 0; i < expected.length; ++i) {
+            assert.equal(emitter.emit.args[i][0], expected[i][0]);
+            assert.equal(emitter.emit.args[i][1], expected[i][1]);
+        }
+    });
+});