From #3559.

I refactored the traversing logic with the decorator pattern.
This modification can conflict with other features, so I separated PR.

----

Before:

```
> node Makefile.js perf

CPU Speed is 3392 with multiplier 7500000
Performance Run #1:  4449.28339ms
Performance Run #2:  4483.965085ms
Performance Run #3:  4434.058325ms
Performance Run #4:  4471.048229ms
Performance Run #5:  4430.7715339999995ms
Performance budget exceeded: 4449.28339ms (limit: 2211.0849056603774ms)
```

After:

```
> node Makefile.js perf

CPU Speed is 3392 with multiplier 7500000
Performance Run #1:  4459.112069ms
Performance Run #2:  4460.901406ms
Performance Run #3:  4406.947205ms
Performance Run #4:  4435.939423ms
Performance Run #5:  4476.988842ms
Performance budget exceeded: 4459.112069ms (limit: 2211.0849056603774ms)
```