@nzakas 2015-09-09
This looks really nice and clean. Just a few comments on the tests. I'd suggest looking into using Sinon instead of creating a fake event emitter.